+++ app/config.py
@@ -33,7 +33,18 @@ def load_config(
     with _config_file_path.open(mode="r", encoding="utf-8") as config_file:
         config_dict = json.load(config_file)
 
-    settings_config = config_dict.get("settings", None)
+    settings_config: dict | None = config_dict.get("settings", None)
+
+    try:
+        number_decimal_places = int(settings_config.get("number_decimal_places", 6))
+        if 0 <= number_decimal_places <= 20:
+            settings_config["number_decimal_places"] = number_decimal_places
+        else:
+            settings_config["number_decimal_places"] = 6
+    except ValueError:
+        settings_config["number_decimal_places"] = 6
+    except TypeError:
+        settings_config["number_decimal_places"] = 6
 
     if "database" in config_dict:
         database_config = config_dict["database"]