+++ app/routers/panelists.py
@@ -32,6 +32,7 @@ from app.models.panelists import PanelistSlug as ModelsPanelistSlug
 router = APIRouter(prefix=f"/v{API_VERSION}/panelists")
 _config = load_config()
 _database_config = _config["database"]
+_settings_config = _config["settings"]
 _database_connection = mysql.connector.connect(**_database_config)
 
 
@@ -181,7 +182,9 @@ async def get_panelists_details():
     """
     try:
         panelist = Panelist(database_connection=_database_connection)
-        panelists = panelist.retrieve_all_details()
+        panelists = panelist.retrieve_all_details(
+            number_decimal_places=_settings_config["number_decimal_places"]
+        )
 
         if panelists:
             return {"panelists": panelists}
@@ -223,7 +226,9 @@ async def get_panelist_details_by_id(
     """
     try:
         panelist = Panelist(database_connection=_database_connection)
-        panelist_details = panelist.retrieve_details_by_id(panelist_id)
+        panelist_details = panelist.retrieve_details_by_id(
+            panelist_id, number_decimal_places=_settings_config["number_decimal_places"]
+        )
 
         if panelist_details:
             return panelist_details
@@ -267,7 +272,9 @@ async def get_random_panelist_details():
     """
     try:
         panelist = Panelist(database_connection=_database_connection)
-        panelist_details = panelist.retrieve_random_details()
+        panelist_details = panelist.retrieve_random_details(
+            number_decimal_places=_settings_config["number_decimal_places"]
+        )
 
         if panelist_details:
             return panelist_details
@@ -313,7 +320,10 @@ async def get_panelist_details_by_slug(
     """
     try:
         panelist = Panelist(database_connection=_database_connection)
-        panelist_details = panelist.retrieve_details_by_slug(panelist_slug.strip())
+        panelist_details = panelist.retrieve_details_by_slug(
+            panelist_slug.strip(),
+            number_decimal_places=_settings_config["number_decimal_places"],
+        )
 
         if panelist_details:
             return panelist_details